sql schema minor change

+++ libs/db/models/content/mixins.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 from sqlalchemy.sql import func
 from sqlalchemy import DateTime
 
@@ -11,13 +11,13 @@ class ContentFeatureMixin:
 
     Schema (current/history):
       - url: varchar (PK only in current; non-null in history)
-      - domain_id: int NOT NULL
+      - domain_id: bigint NOT NULL
       - fetched_at: timestamptz default now()
       - content_length: int default 0
       - content_hash: varchar nullable
       - num_links: int default 0
     """
-    domain_id = Column(Integer, nullable=False)
+    domain_id = Column(BigInteger, nullable=False)
 
     fetched_at = Column(DateTime(timezone=True), server_default=func.now())
