+++ libs/db/models/domain/state.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import BigInteger, Column, Integer, String, Float
 from libs.db.base import Base
 
 class DomainState(Base):
     __tablename__ = "domain_state"
 
-    domain_id = Column(Integer, primary_key=True)
+    domain_id = Column(BigInteger, primary_key=True)
     domain = Column(String, nullable=False, unique=True)
 
     shard_id = Column(Integer, nullable=False)
